Move `ReflectiveFactory` to the `api` module
